+++ DSView/pv/view/trace.h
@@ -84,7 +84,7 @@ public:
     static const QPen SignalAxisPen;
 
 protected:
-    Trace(QString name, int index, int type);
+    Trace(QString name, uint16_t index, int type);
     Trace(QString name, std::list<int> index_list, int type, int sec_index);
 
     /**
